+++ src/routes/user/mfa.ts
@@ -4,19 +4,11 @@ import { ReasonPhrases } from 'http-status-codes';
 
 import { sendError } from '@/errors';
 import { gqlSdk } from '@/utils';
-import { Joi } from '@/validation';
+import { activeMfaType, Joi } from '@/validation';
 
 export const userMfaSchema = Joi.object({
   code: Joi.string().required().description('MFA activation code'),
-  activeMfaType: Joi.alternatives()
-    .try(
-      Joi.string().valid('').empty('').default(null), // accept only empty strings and convert those to null
-      Joi.string().valid('totp')
-    )
-    .example('totp')
-    .description(
-      'Multi-factor authentication type. A null value deactivates MFA'
-    ),
+  activeMfaType,
 }).meta({ className: 'UserMfaSchema' });
 
 export const userMFAHandler: RequestHandler<